+++ src/ofxTLSwitches.h
@@ -62,7 +62,7 @@ class ofxTLSwitches : public ofxTLKeyframes {
     virtual void mouseReleased(ofMouseEventArgs& args, long millis);
     virtual void mouseMoved(ofMouseEventArgs& args, long millis);
     
-    virtual void getSnappingPoints(set<unsigned long>& points);
+    virtual void getSnappingPoints(set<unsigned long long>& points);
     virtual void regionSelected(ofLongRange timeRange, ofRange valueRange);
 
     virtual void unselectAll();
